test: consolidate duplicated test fakes into tests/support/fakes.ts
Extract the jsdom-free fakes that 10 test files re-derived: the
mode-fixture stage stub (byte-identical across five tests/scenes
fixture suites), the synthetic HTMLElement/Document tree shared by the
chrome-slots and chrome-extras suites, and the event-emitting / no-op
presenter controllers used by helpers and presenter-driven. Typing the
chrome fake as the real DOM interfaces removes every as-unknown-as cast
at those call sites; copy-pasted coverage-narration headers trimmed to
one line. Assertions, per-file test counts, and behavior coverage
unchanged.
