+++ tests/scenes/screenshot-rng-fixture.test.ts
@@ -14,74 +14,13 @@
 import { describe, expect, it } from 'vitest';
 import { createSeededRng } from '../../src/runtime/rng';
 import { screenshotRngFixtureScene } from '../../src/scenes/screenshot-rng-fixture';
-
-interface FakeStage {
-  readonly children: { attrs: Map<string, string> }[];
-  readonly element: {
-    setAttribute(name: string, value: string): void;
-    removeAttribute(name: string): void;
-    appendChild(node: unknown): unknown;
-    querySelector(selector: string): {
-      setAttribute(name: string, value: string): void;
-      remove(): void;
-    } | null;
-    ownerDocument: {
-      createElement(tag: string): { setAttribute(name: string, value: string): void };
-    };
-  };
-}
-
-// Stage stub mirroring `scrub-fixture.test.ts`: `appendChild` records
-// children, `querySelector` looks them up by the attribute name in the
-// `[<attr>]` selector, and `ownerDocument.createElement` returns a
-// recording stub. The fixture allocates DOM through the stage's owner
-// document rather than an ambient `document` global (ADR-008 #2).
-const buildStage = (): FakeStage => {
-  const children: { attrs: Map<string, string> }[] = [];
-  return {
-    children,
-    element: {
-      setAttribute: () => undefined,
-      removeAttribute: () => undefined,
-      appendChild: (node: unknown) => {
-        children.push(node as { attrs: Map<string, string> });
-        return node;
-      },
-      querySelector: (selector: string) => {
-        const attr = selector.replace(/^\[|\]$/g, '').split('=')[0];
-        if (attr === undefined) return null;
-        for (let i = 0; i < children.length; i += 1) {
-          const child = children[i];
-          if (child === undefined) continue;
-          if (child.attrs.has(attr)) {
-            return {
-              setAttribute: (name: string, value: string) => child.attrs.set(name, value),
-              remove: () => {
-                children.splice(i, 1);
-              },
-            };
-          }
-        }
-        return null;
-      },
-      ownerDocument: {
-        createElement: () => {
-          const attrs = new Map<string, string>();
-          return {
-            attrs,
-            setAttribute: (name: string, value: string) => attrs.set(name, value),
-          };
-        },
-      },
-    },
-  };
-};
+import { type SceneFixtureStage, buildSceneFixtureStage as buildStage } from '../support/fakes';
 
 const RNG_ATTR = 'data-pulsar-screenshot-rng';
 const TARGET_ATTR = 'data-pulsar-screenshot-rng-target';
 
 // Build a `create`-ready ctx with a seeded `ctx.rng` (PUL-F018).
-const ctxWithSeed = (stage: FakeStage, seed: string): Record<string, unknown> => ({
+const ctxWithSeed = (stage: SceneFixtureStage, seed: string): Record<string, unknown> => ({
   stage: stage.element,
   mode: 'screenshot',
   gsap: { timeline: () => ({}) } as never,
